osa fixes
Fixed typo bit 1 now prints "TS1" instead of "TS2" in the capture OS
types uint8_t → uint16_t for os_types in 5 locations to match the
firmware's UINT16 os_type_mask, enabling Gen6 ERR_OS at bit 8:

+++ inc/switchtec/switchtec.h
@@ -1712,7 +1712,7 @@ struct switchtec_osa_config {
 	uint16_t os_type_lane_mask;
 	uint8_t os_type_direction;
 	uint8_t os_type_link_rate;
-	uint8_t os_type_os_types;
+	uint16_t os_type_os_types;
 
 	/* OS Pattern Trigger */
 	uint16_t os_pat_lane_mask;
@@ -1731,7 +1731,7 @@ struct switchtec_osa_config {
 	uint8_t capture_stop_mode;
 	uint8_t capture_snapshot_mode;
 	uint16_t capture_post_trig_entries;
-	uint8_t capture_os_types;
+	uint16_t capture_os_types;
 };
 
 /**
